refactor(solver): replace MIP page assignment with DP
Wire solve_dp into the assignment phase via PageAssignmentSolver, which now
solves any instance size directly and falls back to a heuristic start solution
when infeasible. Remove the MIP modules, the problem-splitting machinery and
the good_lp/HiGHS dependency. Mark the MIP design doc as superseded.
