+++ docs/design/book_layout_solver_mip/book_layout_solver.md
@@ -1,5 +1,10 @@
 # Book Layout Solver
 
+> **Hinweis:** Die hier beschriebene MIP-Formulierung ist durch den exakten
+> DP-Solver in [`book_layout_solver_dp/`](../book_layout_solver_dp/) abgelöst.
+> Dieses Dokument bleibt als Referenz erhalten, da die Äquivalenz-Argumentation
+> in `book_layout_solver_dp/dp.typ` darauf verweist.
+
 Verteilt eine sortierte Bildsequenz (gruppiert, chronologisch) auf Buchseiten. Zwei Phasen:
 
 1. **MIP-Solver** (`mip/`): Findet eine global optimale Seitenzuteilung unter strukturellen Constraints.
